refactor(craft): replace deadline-timer sleep watcher with SWR poll
The hand-rolled timer (deadline math, retry/in-flight flags, focus and
online listeners) encoded backend reaper timing on the client and had
already cost several review-caught bugs. A 30s SWR poll with a
conditional key detects the same transition with a tenth of the code;
SWR's defaults cover focus revalidation and skip hidden tabs. The
status endpoint slims to {status} and SandboxResponse reverts to its
pre-PR shape — the modal copy no longer claims a specific duration.
