+++ backend/tests/integration/tests/craft/test_sessions_api.py
@@ -132,11 +132,7 @@ def test_get_sandbox_status_returns_status_fields(
     assert response.status_code == 200
     payload = response.json()
     assert "status" in payload
-    assert "last_heartbeat" in payload
-    assert isinstance(payload["idle_timeout_seconds"], int)
-    assert payload["idle_timeout_seconds"] > 0
-    assert isinstance(payload["idle_cleanup_interval_seconds"], int)
-    assert payload["idle_cleanup_interval_seconds"] > 0
+    assert payload["status"] is not None
 
 
 def test_get_sandbox_status_404_for_other_users_session(
